Modification for the WIN_DLL_nanoCLR and WIN32_nanoCLR targets that are not .zip but .dll files. In practive only relevant for the firmware archive.
Bug fix for version that is unknown at the start of the download. The current download assumes that packages with the same version number are identical, so if a package is already present in the cache it is not downloaded again. The original code was unclear: a package was always downloaded, but there was a flag (skipDownload) that intended to bypass the download. The skipDownload was not properly implemented; now it is.

Signed-off-by: Frank Robijn <robijn@good-heavens.nl>